+++ scripts/upgrade-assure/is-linux.go
@@ -1,0 +1,8 @@
+package main
+
+import "runtime"
+
+// is linux?
+func isLinux() bool {
+	return runtime.GOOS == "linux"
+}
